moves nav styles to Navigation.scss
